backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
torch>=2.0.0
transformers==4.53.2
huggingface-hub>=0.20.0
pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0  # ✅ Add this line
python-dotenv>=1.0.0
accelerate>=0.25.0